Start porting dataplane back (#54)
* Add prometheus

* Remove garbage files

* Clean-up

* Start dataplane

* Mark everything as trusted

* Uncomment dependency of the dataplane

* Forgot +gobra

* Going through dataplane imports, check what is necessary

* Start adding more dependencies of the dataplane

* Backup changes

* Backup changes

* WIP

* Backup current status

* Major simplification, sidestep injectivity checks

* Fix CI config for router package

* add metrics

* Add sockctrl

* Add sockctrl to the CI

* Revert changes dataplane.go

+++ pkg/addr/isdas.go
@@ -249,7 +249,7 @@ func (ia IA) IsWildcard() bool {
 //@ decreases
 func (ia IA) String() string {
 	// (VerifiedSCION) Added casts around ia.ISD() and ia.AS() to be able to pass them to 'fmt.Sprintf'
-	return fmt.Sprintf("%d-%s", uint16(ia.ISD()), uint64(ia.AS()))
+	return fmt.Sprintf("%d-%s", ia.ISD(), ia.AS())
 }
 
 // Set implements flag.Value interface
